after naive test

+++ src/pathname.c
@@ -9,14 +9,17 @@ int plain_name(char *path){
 
 
 inode * path_to_inode(char *path, uint dirno){
+    if(strlen(path) == 0) return inode_number_to_inode(1);
     if(plain_name(path))
         return lookup(path, dirno);
-    char *first = path;
+    char pa[100];
+    strncpy(pa, path, strlen(path) + 1);
+    char *first = pa;
     char *rest;
-    for(int i = 0; path[i] != 0; i++){
-        if(path[i] == '/'){
-            path[i] = 0;
-            rest = path + i + 1;
+    for(int i = 0; pa[i] != 0; i++){
+        if(pa[i] == '/'){
+            pa[i] = 0;
+            rest = pa + i + 1;
             break;
         }
     }
@@ -56,12 +59,14 @@ void next_dir(char *path, uint dirno, uint type){
         if(type == 3) createsoft(path, dirno);
         return ;
     }
-    char *first = path;
+    char pa[100];
+    strncpy(pa, path, strlen(path) + 1);
+    char *first = pa;
     char *rest;
-    for(int i = 0; path[i] != 0; i++){
-        if(path[i] == '/'){
-            path[i] = 0;
-            rest = path + i + 1;
+    for(int i = 0; pa[i] != 0; i++){
+        if(pa[i] == '/'){
+            pa[i] = 0;
+            rest = pa + i + 1;
             break;
         }
     }
@@ -70,20 +75,21 @@ void next_dir(char *path, uint dirno, uint type){
     next_dir(rest, newdirno, type);
 }
 
-void mkdir(char *path){
-    inode *ino;
-    if(*path == '/')
-        next_dir(path + 1, rootdirno, 1);
-    else
-        next_dir(path, cwd.inum, 1);
-}
-void touch(char *path, uint type){
-    inode *ino;
+void mkfile(char *path, uint type){
     if(*path == '/')
         next_dir(path + 1, rootdirno, type);
     else
         next_dir(path, cwd.inum, type);
 }
+
+void mkdir(char *path){
+    mkfile(path, 1);
+}
+
+void touch(char *path){
+    mkfile(path, 2);
+}
+
 
 void handle_linkname(char *path, uint dirno, uint inum, int flag){
     if(plain_name(path)){
@@ -91,12 +97,14 @@ void handle_linkname(char *path, uint dirno, uint inum, int flag){
         else if(flag == 1) deletelink(inode_number_to_inode(dirno), path);
         return ;
     }
-    char *first = path;
+    char pa[100];
+    strncpy(pa, path, strlen(path) + 1);
+    char *first = pa;
     char *rest;
-    for(int i = 0; path[i] != 0; i++){
-        if(path[i] == '/'){
-            path[i] = 0;
-            rest = path + i + 1;
+    for(int i = 0; pa[i] != 0; i++){
+        if(pa[i] == '/'){
+            pa[i] = 0;
+            rest = pa + i + 1;
             break;
         }
     }
@@ -118,7 +126,11 @@ void link(char *path, char *linkname){
         return ;
     }
     if(ino->type == 1){
-        printf("can't link dir");
+        printf("can't link dir\n");
+        return;
+    }
+    if(ino->type == 3){
+        printf("can't link soft link\n");
         return;
     }
     if(*linkname == '/') 
@@ -132,13 +144,16 @@ void unlink(char *path){
     inode *ino;
     if(*path == '/'){
         ino = path_to_inode(path + 1, rootdirno);
-        handle_linkname(path + 1, rootdirno, 0, 1);
     }
     else{
         ino = path_to_inode(path, cwd.inum);
-        handle_linkname(path, cwd.inum, 0, 1);
     }
     if(ino != NULL){
+        //printf("%d\n", ino->inum);
+        if(*path == '/')
+            handle_linkname(path + 1, rootdirno, 0, 1);
+        else
+            handle_linkname(path, cwd.inum, 0, 1);
         ino->ref--;
         if(ino->ref == 0){
             ifree(ino);
@@ -155,49 +170,87 @@ void m_remove(char *path){
         return;
     }
     inode *ino;
-    if(*path){
+    if(*path == '/'){
         ino = path_to_inode(path + 1, rootdirno);
     }
     else{
         ino = path_to_inode(path, cwd.inum);
     }
+    if(ino == NULL){
+        printf("not found this path\n");
+        return;
+    }
     if(ino->type == 1){
+        //printf("1");
         if(dirempty(ino)){
             unlink(path);
-            printf("remove dir successful");
+            printf("remove dir %s successful\n", path);
+        }
+        else{
+            printf("dir %s not empty, can't remove\n", path);
+            return;
         }
     }
     else{
         unlink(path);
-        printf("remove file successful");
+        printf("remove file successful\n");
     }
 }
 
 void m_rename(char *name1, char *name2){
     int i, j;
-    for(i = strlen(name1) - 1; i >= 0; i--){
-        if(name1[i] == '/') break;
+    char n1[100], n2[100];
+    if(*name1 != '/'){
+        strncpy(n1, pwd, strlen(pwd) + 1);
+        int len = strlen(n1);
+        if(n1[len - 1] != '/'){
+            n1[len - 1] = '/';
+            n1[len] = 0;
+        }
+        strncat(n1, name1, strlen(name1));
     }
-    for(j = strlen(name2) - 1; j >= 0; j--){
-        if(name2[j] == '/') break;
+    else{
+        strncpy(n1, name1, strlen(name1) + 1);
     }
-    if(i == j && strncmp(name1, name2, i) == 0){
-        char *s1 = name1 + i + 1;
-        char *s2 = name2 + j + 1;
-        name1[i] = 0;
+    if(*name2 != '/'){
+        strncpy(n2, pwd, strlen(pwd) + 1);
+        int len = strlen(n2);
+        if(n1[len - 1] != '/'){
+            n1[len - 1] = '/';
+            n1[len] = 0;
+        }
+        strncat(n2, name2, strlen(name2));
+    }
+    else{
+        strncpy(n2, name2, strlen(name2) + 1);
+    }
+    for(i = strlen(n1) - 1; i >= 0; i--){
+        if(n1[i] == '/') break;
+    }
+    for(j = strlen(n2) - 1; j >= 0; j--){
+        if(n2[j] == '/') break;
+    }
+    if(i == j && strncmp(n1, n2, i) == 0){
+        char *s1 = n1 + i + 1;
+        char *s2 = n2 + j + 1;
+        n1[i] = 0;
         inode *ino;
-        if(*name1 == '/')
+        if(*n1 == '/')
             ino = path_to_inode(name1 + 1, rootdirno);
-        else    
+        else if(strlen(n1) > 0)  
             ino = path_to_inode(name1, cwd.inum);
+        else{
+            ino = inode_number_to_inode(1);
+        }
         if(ino != NULL){
             if(ino->type != 1){
                 printf("not a dir\n");
                 return ;
             }
             else{
+                int off;
                 dirent de;
-                for(int off = 0; off < ino->size; off += sizeof(dirent)){
+                for(off = 0; off < ino->size; off += sizeof(dirent)){
                     if(readinode(ino, (char*)&de, off, sizeof(dirent)) != sizeof(dirent)){
                         printf("dirlink read");
                         return ;
@@ -206,8 +259,12 @@ void m_rename(char *name1, char *name2){
                         break;
                     if(strncmp(de.name, s1, DSIZE) == 0){
                         strncpy(de.name, s2, DSIZE);
-                        return;
+                        break;
                     }
+                }
+                if(writeinode(ino, (char *)&de, off, sizeof(dirent))!= sizeof(dirent)){
+                    printf("dirlink write wrong\n");
+                    return ;
                 }
             }
         }
